src/holopy/quantum/error_correction.py
"""
Quantum error correction system with holographic stabilizers.
"""
from typing import List, Tuple, Optional, Dict
import numpy as np
from scipy.sparse import csr_matrix
import logging
from dataclasses import dataclass
from ..config.constants import (
    INFORMATION_GENERATION_RATE,
    PLANCK_CONSTANT,
    SPEED_OF_LIGHT
)

logger = logging.getLogger(__name__)

@dataclass
class StabilizerMetrics:
    """Metrics for stabilizer code performance."""
    syndrome_pattern: np.ndarray
    correction_fidelity: float
    logical_error_rate: float
    code_distance: int
    recovery_time: float

class HolographicStabilizer:
    """Implements holographic quantum error correction."""
    
    def __init__(
        self,
        code_distance: int,
        spatial_points: int,
        recovery_threshold: float = 0.95
    ):
        """
        Initialize holographic stabilizer code.
        
        Args:
            code_distance: Distance of the quantum error correcting code
            spatial_points: Number of spatial points
            recovery_threshold: Threshold for error recovery
        """
        self.code_distance = code_distance
        self.spatial_points = spatial_points
        self.recovery_threshold = recovery_threshold
        
        # Initialize stabilizer generators and logical operators
        self._initialize_code()
        
        logger.info(
            f"Initialized HolographicStabilizer with distance {code_distance}"
        )
    
    def _initialize_code(self) -> None:
        """Initialize stabilizer code structure."""
        try:
            # Create stabilizer generators
            self.stabilizers = self._construct_stabilizers()
            
            # Create logical operators
            self.logical_x = self._construct_logical_x()
            self.logical_z = self._construct_logical_z()
            
            # Initialize syndrome table
            self._build_syndrome_table()
            
            logger.debug("Initialized stabilizer code structure")
            
        except Exception as e:
            logger.error(f"Code initialization failed: {str(e)}")
            raise
    
    def _construct_stabilizers(self) -> List[csr_matrix]:
        """Construct stabilizer generators with holographic constraints."""
        try:
            stabilizers = []
            n_stabilizers = self.code_distance**2 - 1
            
            for i in range(n_stabilizers):
                # Create plaquette operator
                plaquette = self._create_plaquette_operator(i)
                
                # Apply holographic corrections
                plaquette = self._apply_holographic_corrections(plaquette)
                
                stabilizers.append(plaquette)
            
            return stabilizers
            
        except Exception as e:
            logger.error(f"Stabilizer construction failed: {str(e)}")
            raise
    
    def _create_plaquette_operator(self, index: int) -> csr_matrix:
        """Create individual plaquette operator."""
        try:
            # Calculate plaquette position
            row = index // self.code_distance
            col = index % self.code_distance
            
            # Create sparse operator
            data = []
            rows = []
            cols = []
            
            # Add X and Z terms with phase factors
            for i in range(4):
                pos = self._get_vertex_position(row, col, i)
                phase = np.exp(2j * np.pi * i / 4)
                
                data.append(phase)
                rows.append(pos)
                cols.append(pos)
            
            return csr_matrix(
                (data, (rows, cols)),
                shape=(self.spatial_points, self.spatial_points)
            )
            
        except Exception as e:
            logger.error(f"Plaquette creation failed: {str(e)}")
            raise
    
    def _apply_holographic_corrections(
        self,
        operator: csr_matrix
    ) -> csr_matrix:
        """Apply holographic corrections to stabilizer operators."""
        try:
            # Apply information rate decay
            decay_factor = np.exp(-INFORMATION_GENERATION_RATE * 
                                operator.diagonal().mean())
            
            # Apply holographic bound
            max_weight = np.log2(self.spatial_points)
            weight = np.sum(np.abs(operator.data))
            
            if weight > max_weight:
                operator.data *= max_weight / weight
            
            return operator * decay_factor
            
        except Exception as e:
            logger.error(f"Holographic correction failed: {str(e)}")
            raise
    
    def measure_syndrome(
        self,
        state: np.ndarray
    ) -> Tuple[np.ndarray, StabilizerMetrics]:
        """
        Measure error syndrome with holographic constraints.
        
        Args:
            state: Quantum state to measure
            
        Returns:
            Tuple of (syndrome_pattern, stabilizer_metrics)
        """
        try:
            # Initialize syndrome pattern
            syndrome = np.zeros(len(self.stabilizers), dtype=bool)
            
            # Measure each stabilizer
            for i, stabilizer in enumerate(self.stabilizers):
                # Calculate expectation value
                expectation = np.abs(
                    np.vdot(state, stabilizer @ state)
                )
                
                # Apply measurement with holographic noise
                noise = np.random.normal(
                    0,
                    INFORMATION_GENERATION_RATE
                )
                syndrome[i] = expectation + noise < self.recovery_threshold
            
            # Calculate metrics
            metrics = StabilizerMetrics(
                syndrome_pattern=syndrome,
                correction_fidelity=self._calculate_fidelity(syndrome),
                logical_error_rate=self._estimate_logical_error_rate(syndrome),
                code_distance=self.code_distance,
                recovery_time=self._estimate_recovery_time(syndrome)
            )
            
            logger.debug(f"Measured syndrome pattern: {syndrome}")
            
            return syndrome, metrics
            
        except Exception as e:
            logger.error(f"Syndrome measurement failed: {str(e)}")
            raise
    
    def apply_correction(
        self,
        state: np.ndarray,
        syndrome: np.ndarray
    ) -> Tuple[np.ndarray, float]:
        """
        Apply error correction based on syndrome measurement.
        
        Args:
            state: State to correct
            syndrome: Measured error syndrome
            
        Returns:
            Tuple of (corrected_state, correction_fidelity)
        """
        try:
            # Look up correction operator
            correction = self._get_correction_operator(syndrome)
            
            # Apply correction with holographic constraints
            corrected_state = correction @ state
            
            # Normalize
            corrected_state /= np.sqrt(np.sum(np.abs(corrected_state)**2))
            
            # Calculate correction fidelity
            fidelity = np.abs(np.vdot(corrected_state, state))
            
            logger.debug(f"Applied correction with fidelity {fidelity:.4f}")
            
            return corrected_state, fidelity
            
        except Exception as e:
            logger.error(f"Error correction failed: {str(e)}")
            raise
    
    def _calculate_fidelity(self, syndrome: np.ndarray) -> float:
        """Calculate correction fidelity from syndrome pattern."""
        try:
            # Calculate error weight
            error_weight = np.sum(syndrome)
            
            # Estimate fidelity based on code distance
            return np.exp(-error_weight / self.code_distance)
            
        except Exception as e:
            logger.error(f"Fidelity calculation failed: {str(e)}")
            raise
    
    def _estimate_logical_error_rate(self, syndrome: np.ndarray) -> float:
        """Estimate logical error rate from syndrome pattern."""
        try:
            # Calculate error chains
            chains = self._find_error_chains(syndrome)
            
            # Estimate logical error probability
            p_logical = 1 - np.exp(-len(chains) / self.code_distance)
            
            return p_logical
            
        except Exception as e:
            logger.error(f"Error rate estimation failed: {str(e)}")
            raise 